refactor(app): extract ChartPane from QuantickApp
Split the ~6800-line QuantickApp in two along the line the split view and
the workspace tabs will need (docs/ux/ui-design-model.md §11). The new
crate::pane::ChartPane owns what answers "what is on this canvas": the
bar series and its selector state, the viewport and price scale, the
cached geometry the input pass reads before the draw recomputes it, the
drawings anchored to bar indices, the orderflow view and the indicator
slots with their worker.

The window keeps what feeds and frames that canvas: feed channels,
connection state, notices, history counters, loading, replay link, book
capture epoch, and all app chrome (config, style, timezone, dock,
toolrail, status bar, replay browser, inspector, frame metrics).
slot_kinds stays with the persistence layer it belongs to — one state
file records what the window had open, and a per-pane copy would need a
key no pane has yet.

handle_navigation and draw_chart, and every helper that only touched pane
fields, become ChartPane methods taking PaneInput (the shared toolbox and
preset store) or PaneRender (style, timezone, armed tool, symbol) instead
of reaching into the app. Their egui interaction ids — chart_nav,
time_nav, lane_time_nav, price_nav, lane_divider, drawing_placement — are
now derived from ChartPane::id so two panes cannot share a drag.

Behaviour is unchanged: one pane, same layout, same gestures. plot_split
keeps its home and its tests.
